Move tasks to manage TariffPrices and NCSOConcessions out of dmd app
