+++ openprescribing/pipeline/management/commands/reconcile_ncso_concessions.py
@@ -3,7 +3,8 @@ from __future__ import print_function
 from django.core.management import BaseCommand
 
 from openprescribing.utils import get_input
-from dmd.models import NCSOConcession, DMDVmpp
+from dmd2.models import VMPP
+from frontend.models import NCSOConcession
 from gcutils.bigquery import Client
 
 
@@ -30,7 +31,7 @@ class Command(BaseCommand):
             q = get_input()
             self.stdout.write('')
 
-            candidates = DMDVmpp.objects.filter(nm__icontains=q).order_by('nm')
+            candidates = VMPP.objects.filter(nm__icontains=q).order_by('nm')
 
             num_candidates = candidates.count()
